# www/public/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "description": "Issued when a browser pairs with the daemon.",
        "in": "cookie",
        "name": "picode_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The HTTP API the PiCode web UI itself speaks — same daemon, same pairing-gated session. Generated from the server's route registration by cmd/picode-openapi; do not edit the JSON by hand. Authenticate with the picode_session cookie issued when you pair a browser (Settings → Pairing).",
    "title": "PiCode HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/inbox": {
      "delete": {
        "operationId": "deleteApiInbox",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "inbox"
        ]
      },
      "get": {
        "operationId": "getApiInbox",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "inbox"
        ]
      },
      "post": {
        "operationId": "postApiInbox",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/inbox/{id}": {
      "delete": {
        "operationId": "deleteApiInbox",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/inbox/{id}/respond": {
      "post": {
        "operationId": "postApiInboxRespond",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/inbox/{id}/state": {
      "post": {
        "operationId": "postApiInboxState",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/llama": {
      "get": {
        "operationId": "getApiLlama",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "llama"
        ]
      }
    },
    "/api/llama/download": {
      "post": {
        "operationId": "postApiLlamaDownload",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "llama"
        ]
      }
    },
    "/api/llama/hf": {
      "get": {
        "operationId": "getApiLlamaHf",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "llama"
        ]
      }
    },
    "/api/llama/hf/info": {
      "get": {
        "operationId": "getApiLlamaHfInfo",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "llama"
        ]
      }
    },
    "/api/llama/load": {
      "post": {
        "operationId": "postApiLlamaLoad",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "llama"
        ]
      }
    },
    "/api/llama/unload": {
      "post": {
        "operationId": "postApiLlamaUnload",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "llama"
        ]
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "description": "local daemon",
      "url": "http://127.0.0.1:8445"
    }
  ],
...
}